requirements.txt
asttokens @ file:///build/asttokens-2.2.1/dist/asttokens-2.2.1-py2.py3-none-any.whl
attrs==23.2.0
Babel==2.9.1
backcall @ file:///build/backcall-0.2.0/dist/backcall-0.2.0-py3-none-any.whl
beautifulsoup4==4.12.3
cached-property==1.5.2
certifi==2024.2.2
cffi==1.16.0
cfgv @ file:///build/cfgv-3.3.1/dist/cfgv-3.3.1-py2.py3-none-any.whl
chardet==4.0.0
cowsay==6.1
cryptography==3.4.8
debugpy @ file:///build/source/dist/debugpy-1.6.7-cp310-cp310-linux_x86_64.whl
decorator @ file:///build/decorator-5.1.1/dist/decorator-5.1.1-py3-none-any.whl
distlib @ file:///build/distlib-0.3.6/dist/distlib-0.3.6-py2.py3-none-any.whl
docopt==0.6.2
docutils @ file:///build/docutils-0.19/dist/docutils-0.19-py3-none-any.whl
ebaysdk==2.1.5
executing @ file:///build/source/dist/executing-1.2.0-py2.py3-none-any.whl
filelock @ file:///build/filelock-3.9.0/dist/filelock-3.9.0-py3-none-any.whl
freezegun==1.1.0
geoip2==2.9.0
gevent==21.8.0
greenlet==1.1.2
identify @ file:///build/source/dist/identify-2.5.24-py2.py3-none-any.whl
idna==2.10
ipdb @ file:///build/ipdb-0.13.13/dist/ipdb-0.13.13-py3-none-any.whl
ipython @ file:///build/ipython-8.11.0/dist/ipython-8.11.0-py3-none-any.whl
ipython-genutils @ file:///build/ipython_genutils-0.2.0/dist/ipython_genutils-0.2.0-py2.py3-none-any.whl
isodate==0.6.1
jedi @ file:///build/source/dist/jedi-0.18.2-py2.py3-none-any.whl
Jinja2==3.0.3
libsass @ file:///build/source/dist/libsass-0.22.0-cp310-abi3-linux_x86_64.whl
lxml==4.8.0
MarkupSafe==2.0.1
matplotlib-inline @ file:///build/matplotlib-inline-0.1.6/dist/matplotlib_inline-0.1.6-py3-none-any.whl
maxminddb==2.6.0
nodeenv @ file:///build/source/dist/nodeenv-1.8.0-py2.py3-none-any.whl
num2words==0.5.10
odoo-client-lib==1.2.2
odoo-stubs==0.1.dev0
ofxparse==0.21
parso @ file:///build/parso-0.8.3/dist/parso-0.8.3-py2.py3-none-any.whl
passlib==1.7.4
path @ file:///build/path-16.6.0/dist/path-16.6.0-py3-none-any.whl
pexpect @ file:///build/pexpect-4.8.0/dist/pexpect-4.8.0-py2.py3-none-any.whl
pickleshare @ file:///build/pickleshare-0.7.5/dist/pickleshare-0.7.5-py2.py3-none-any.whl
Pillow==9.0.1
platformdirs @ file:///build/source/dist/platformdirs-3.0.0-py3-none-any.whl
polib==1.1.1
pre-commit @ file:///build/source/dist/pre_commit-3.3.2-py2.py3-none-any.whl
prompt-toolkit @ file:///build/prompt_toolkit-3.0.38/dist/prompt_toolkit-3.0.38-py3-none-any.whl
psutil==5.9.0
psycopg2==2.9.2
ptyprocess @ file:///build/ptyprocess-0.7.0/dist/ptyprocess-0.7.0-py3-none-any.whl
pure-eval @ file:///build/source/dist/pure_eval-0.2.2-py3-none-any.whl
py==1.11.0
pyasn1 @ file:///build/pyasn1-0.5.0/dist/pyasn1-0.5.0-py2.py3-none-any.whl
pyasn1-modules @ file:///build/source/dist/pyasn1_modules-0.3.0-py2.py3-none-any.whl
pycparser==2.22
pydevd-odoo==1.2.2
pydot==1.4.2
Pygments @ file:///build/Pygments-2.14.0/dist/Pygments-2.14.0-py3-none-any.whl
pyOpenSSL==21.0.0
pyparsing==3.1.2
PyPDF2==1.26.0
pyserial==3.5
python-dateutil==2.8.1
python-ldap @ file:///build/source/dist/python_ldap-3.4.4-cp310-cp310-linux_x86_64.whl
python-stdnum==1.17
pytz==2024.1
pyusb==1.2.1
PyYAML @ file:///build/source/dist/PyYAML-6.0-cp310-cp310-linux_x86_64.whl
qrcode==7.3.1
reportlab==3.6.8
requests==2.25.1
requests-file==2.0.0
requests-toolbelt==1.0.0
rjsmin==1.1.0
six @ file:///build/six-1.16.0/dist/six-1.16.0-py2.py3-none-any.whl
soupsieve==2.5
stack-data @ file:///build/source/dist/stack_data-0.2.0-py3-none-any.whl
toml @ file:///build/toml-0.10.2/dist/toml-0.10.2-py2.py3-none-any.whl
tomli @ file:///build/source/dist/tomli-2.0.1-py3-none-any.whl
traitlets @ file:///build/traitlets-5.9.0/dist/traitlets-5.9.0-py3-none-any.whl
urllib3==1.26.5
virtualenv @ file:///build/virtualenv-20.19.0/dist/virtualenv-20.19.0-py3-none-any.whl
vobject==0.9.6.1
wcwidth @ file:///build/wcwidth-0.2.6/dist/wcwidth-0.2.6-py2.py3-none-any.whl
Werkzeug==2.0.2
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
zeep==4.1.0
zope.event==5.0
zope.interface==6.2
